Comp push

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Config/Commands/Custom/DynamicWaitCommand.java
@@ -8,12 +8,12 @@ import java.util.function.DoubleSupplier;
 public class DynamicWaitCommand extends CommandBase {
     private final DoubleSupplier timeSupplier;
     private long startTime, waitTime;
-    private int multiplier = 1;
+    private double multiplier = 1;
 
     public DynamicWaitCommand(DoubleSupplier supplier) {
         timeSupplier = supplier;
     }
-    public DynamicWaitCommand(DoubleSupplier supplier, int multipler) {
+    public DynamicWaitCommand(DoubleSupplier supplier, double multipler) {
         timeSupplier = supplier;
         this.multiplier = multipler;
     }
@@ -22,7 +22,7 @@ public class DynamicWaitCommand extends CommandBase {
     @Override
     public void initialize() {
         startTime = System.currentTimeMillis();
-        waitTime = (long) timeSupplier.getAsDouble() * multiplier; // ✔ updated each time it runs
+        waitTime = (long) (timeSupplier.getAsDouble() * multiplier); // ✔ updated each time it runs
 
     }
